fix(menu): fix/menu-item-group-not-working
- Make the menu-item-group exported correctly
