+++ packages/menu/__tests__/menu.spec.ts
@@ -3,7 +3,7 @@ import { nextTick } from 'vue'
 import { sleep } from '@element-plus/test-utils'
 
 import Menu from '../src/menu.vue'
-import MenuGroup from '../src/menuGroup.vue'
+import MenuGroup from '../src/menuItemGroup.vue'
 import MenuItem from '../src/menuItem.vue'
 import SubMenu from '../src/submenu.vue'
